* Fixed: More Compiler warnings (Unsigned/Signed comparison).
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@4031 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/debuggergdb/debuggeroptionsprjdlg.cpp
@@ -124,7 +124,7 @@ void DebuggerOptionsProjectDlg::OnApply()
     wxListBox* control = XRCCTRL(*this, "lstSearchDirs", wxListBox);
 
     m_OldPaths.Clear();
-    for (int i = 0; i < control->GetCount(); ++i)
+    for (int i = 0; i < (int)control->GetCount(); ++i)
     {
         m_OldPaths.Add(control->GetString(i));
     }
